Set project output type to Library
Updated the `SuperMarket.API.csproj` file to set the `<OutputType>` property to `Library`, enabling the project to produce a DLL instead of an executable. This change prepares the project to be used as a dependency in other projects. No changes were made to the target framework, nullable reference type settings, or existing package references.